Move newsletter pages to app router
